feature/OSDOP-260/remove-users-from-team (#94)
* Alternative pattern for team management

* Fixed failing test

* Fixed several PR comments

* Converted hard-coded ID to use actual user object for teams repositories

* Added redirect if team has the incorrect team type for the endpoint

* Addressed several PR comments

* Added tests for teams package

* Removed Team Type and Team screens

* Removed Team Type and Team screens

* Added test for getUserViewsForTeam

* Added more tests and removed redirect endpoint

- Renamed userId to wuaId

* Fixed several PR comments

* Exclude title from TeamMemberView display name when null

* Added RegulatorTeamService

* Moved regulator management files to regulator sub package. Changed response code to forbidden if user is not part of a regulator team.

* Addressed PR cleanup comments

* Restructured team members table to include roles

* Fixed mailTo template

* Added distinct to TeamRepository and a test to verify behaviour

* Added util test builders for team and team member and refactored database integration test to use

* added security annotation to restrict endpoints to users who are members of the team associated

* added security annotation to restrict regulator add member endpoints to access managers only

* Added ability for team members to be shown on screen on the team member page

* Added tests for TeamMemberViewService getTeamMemberViewsForTeam method

* Rebased onto OSDOP-42/7

* Fixed issues after rebase

* Addressed PR comments

* Simplified getTeamMember method

* Removed file added during rebase

* Moved default TeamMemberViewTestUtil properties initialisation

+++ src/main/java/uk/co/nstauthority/offshoresafetydirective/energyportal/user/EnergyPortalUserDto.java
@@ -1,7 +1,7 @@
 package uk.co.nstauthority.offshoresafetydirective.energyportal.user;
 
 public record EnergyPortalUserDto(
-    int webUserAccountId,
+    long webUserAccountId,
     String title,
     String forename,
     String surname,
